// Downloads/dpm-ui-web-and-mobile-development/web-apps/consumer-portal/src/types/quoteAndBuy.ts
  export interface SuccessPagePolicyData {
    policyNo: string;
    startDate: string;
    expiryDate: string;
    coverageName: string;
    customerNameEnglish: string;
    customerNameArabic: string;
    nationalityId: string;
    mobileNo: string;
    premiumAmount: string;
    sumInsured: string;
    repairCondition: string;
    vehicleMake: string | undefined;
    [key: string]: string | undefined;
  }